backend/requirements.txt
# Django & DRF
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9

# CORS
django-cors-headers==4.3.1

# File handling
Pillow==10.1.0

# Utilities
requests==2.31
python-decouple==3.8

# PDF & DOCX Parsing
PyPDF2==3.0.1                    
python-docx==1.1.0               
pdfplumber==0.10.3              

# NLP & Text Processing
spacy==3.7.2                     

# Alternative/Additional NLP tools
transformers==4.36.0
nltk==3.8.1

# Pattern Matching
regex==2023.12.25               
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1

# Date Parsing
python-dateutil==2.8.2
dateparser==1.2.0

# File Validation
python-magic==0.4.27
Pillow==10.1.0

# Task Queue
celery==5.4.0
redis==5.0.1
django-celery-beat==2.6.0
django-celery-results==2.5.1
